introduced custom annotation to ease temp folder cleanup after unit tests
