+++ src/test/java/com/github/stickerifier/stickerify/ResourceHelper.java
@@ -7,8 +7,6 @@ import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 
 public final class ResourceHelper {
 
@@ -36,28 +34,5 @@ public final class ResourceHelper {
 		assumeTrue(resource != null, "Test resource [%s] not found.".formatted(filename));
 
 		return new File(resource.getFile());
-	}
-
-	public static void deleteTempFiles() throws IOException {
-		try (var files = Files.list(Path.of(System.getProperty("java.io.tmpdir")))) {
-			files.filter(Files::isRegularFile)
-					.map(Path::toFile)
-					.filter(ResourceHelper::stickerifyFiles)
-					.forEach(ResourceHelper::deleteFile);
-		}
-	}
-
-	private static boolean stickerifyFiles(File file) {
-		var fileName = file.getName();
-
-		return fileName.startsWith("Stickerify") || fileName.startsWith("OriginalFile");
-	}
-
-	private static void deleteFile(File file) {
-		try {
-			Files.delete(file.toPath());
-		} catch (IOException e) {
-			abort("The file could not be deleted from the system.");
-		}
 	}
 }
