+++ src/test/java/com/github/stickerifier/stickerify/junit/TempFilesCleanerExtension.java
@@ -1,0 +1,46 @@
+package com.github.stickerifier.stickerify.junit;
+
+import static org.junit.jupiter.api.Assumptions.abort;
+
+import org.junit.jupiter.api.extension.AfterAllCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+/**
+ * Custom JUnit extension used to clear temp files
+ * generated by unit tests during their execution.
+ */
+public class TempFilesCleanerExtension implements AfterAllCallback {
+
+	@Override
+	public void afterAll(ExtensionContext context) throws IOException {
+		deleteTempFiles();
+	}
+
+	private static void deleteTempFiles() throws IOException {
+		try (var files = Files.list(Path.of(System.getProperty("java.io.tmpdir")))) {
+			files.filter(Files::isRegularFile)
+					.map(Path::toFile)
+					.filter(TempFilesCleanerExtension::stickerifyFiles)
+					.forEach(TempFilesCleanerExtension::deleteFile);
+		}
+	}
+
+	private static boolean stickerifyFiles(File file) {
+		var fileName = file.getName();
+
+		return fileName.startsWith("Stickerify-") || fileName.startsWith("OriginalFile-");
+	}
+
+	private static void deleteFile(File file) {
+		try {
+			Files.delete(file.toPath());
+		} catch (IOException e) {
+			abort("The file could not be deleted from the system.");
+		}
+	}
+}
